Announce: Add now and list subcommands, refactor and update permissions for types
